# openapi/wsapi-api.yml
openapi: 3.0.3
info:
  title: WSAPI - WhatsApp API
  description: RESTful HTTP API for WhatsApp via whatsmeow. Supports multi-instance connections, webhook event delivery, and HMAC-SHA256 signing.
  version: 2.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
...
security:
  - InstanceAuth: []
paths:
...
  /messages/text:
    post:
      tags: [Messages]
      summary: Send a text message
      description: Send a text message to a WhatsApp contact or group. The response contains the message ID, which can be used later for reply, edit, delete, star, and other actions.
      operationId: sendText
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendTextRequest"
            examples:
              simpleMessage:
                summary: Plain text message
                value:
                  to: "1234567890@s.whatsapp.net"
                  text: "Hello world!"
              withMentions:
                summary: Text with user mention
                value:
                  to: "1234567890@g.us"
                  text: "Hi @1234567890"
                  mentions: ["1234567890@s.whatsapp.net"]
              reply:
                summary: Reply to a previous message
                value:
                  to: "1234567890@s.whatsapp.net"
                  text: "On it!"
                  replyTo: "01234567890123456"
                  replyToSenderId: "1234567890:12@s.whatsapp.net"
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /messages/image:
    post:
      tags: [Messages]
      summary: Send an image message
      description: Send an image message via URL or base64-encoded data. The response contains the message ID.
      operationId: sendImage
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMediaRequest"
            examples:
              byURL:
                summary: Image by URL
                value:
                  to: "1234567890@s.whatsapp.net"
                  url: "https://example.com/cat.jpg"
                  mimeType: "image/jpeg"
                  caption: "Cat pic"
              byBase64ViewOnce:
                summary: Image by base64 sent as view-once
                value:
                  to: "1234567890@g.us"
                  data: "/9j/4AAQSkZJRgABAQ..."
                  mimeType: "image/png"
                  caption: "One-time secret"
                  viewOnce: true
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /messages/video:
    post:
      tags: [Messages]
      summary: Send a video message
      description: Send a video message via URL or base64-encoded data. The response contains the message ID.
      operationId: sendVideo
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMediaRequest"
            examples:
              byURL:
                summary: Video by URL with caption
                value:
                  to: "1234567890@s.whatsapp.net"
                  url: "https://example.com/clip.mp4"
                  mimeType: "video/mp4"
                  caption: "Watch this"
              byBase64ViewOnce:
                summary: Video by base64 as view-once
                value:
                  to: "1234567890@g.us"
                  data: "AAAAGGZ0eXBtcDQy..."
                  mimeType: "video/mp4"
                  viewOnce: true
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /messages/audio:
    post:
      tags: [Messages]
      summary: Send an audio message
      description: Send an audio message via URL or base64-encoded data. The response contains the message ID.
      operationId: sendAudio
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMediaRequest"
            examples:
              byURL:
                summary: Audio by URL
                value:
                  to: "1234567890@s.whatsapp.net"
                  url: "https://example.com/audio.mp3"
                  mimeType: "audio/mpeg"
              byBase64ViewOnce:
                summary: Audio by base64 as view-once
                value:
                  to: "1234567890@g.us"
                  data: "SUQzBAAAAAAAI1RTU0U..."
                  mimeType: "audio/mp3"
                  viewOnce: true
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /messages/voice:
    post:
      tags: [Messages]
      summary: Send a voice message
      description: Send a voice (push-to-talk) message via URL or base64-encoded data. The response contains the message ID.
      operationId: sendVoice
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMediaRequest"
            examples:
              byURL:
                summary: Voice by URL
                value:
                  to: "1234567890@s.whatsapp.net"
                  url: "https://example.com/ptt.ogg"
              byBase64:
                summary: Voice by base64
                value:
                  to: "1234567890@g.us"
                  data: "T2dnUwACAAAAAAAAAAA..."
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /messages/document:
    post:
      tags: [Messages]
      summary: Send a document message
      description: Send a document via URL or base64-encoded data. The `filename` field is required. The response contains the message ID.
      operationId: sendDocument
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendDocumentRequest"
            examples:
              byURL:
                summary: Document by URL with caption
                value:
                  to: "1234567890@s.whatsapp.net"
                  url: "https://example.com/report.pdf"
                  filename: "report.pdf"
                  caption: "Monthly report"
              byBase64:
                summary: Document by base64
                value:
                  to: "1234567890@g.us"
                  data: "JVBERi0xLjQKJeLjz9M..."
                  filename: "numbers.xlsx"
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /messages/sticker:
    post:
      tags: [Messages]
      summary: Send a sticker message
      description: Send a sticker (WebP format) via URL or base64-encoded data. Use `isAnimated` for animated stickers. The response contains the message ID.
      operationId: sendSticker
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendStickerRequest"
            examples:
              byURL:
                summary: Static sticker by URL
                value:
                  to: "1234567890@s.whatsapp.net"
                  url: "https://example.com/sticker.webp"
                  isAnimated: false
              byBase64Animated:
                summary: Animated sticker by base64
                value:
                  to: "1234567890@g.us"
                  data: "UklGRiIAAABXRUJQVlA4..."
                  isAnimated: true
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /messages/contact:
    post:
      tags: [Messages]
      summary: Send a contact card
      description: Send a contact card via display name (auto-generated vCard) or a raw vCard string. The response contains the message ID.
      operationId: sendContact
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendContactRequest"
            examples:
              withDisplayName:
                summary: Contact with display name
                value:
                  to: "1234567890@s.whatsapp.net"
                  displayName: "Jane Doe"
              withVCard:
                summary: Contact using vCard
                value:
                  to: "1234567890@g.us"
                  vcard: "BEGIN:VCARD\nVERSION:3.0\nN:Doe;Jane;;;\nFN:Jane Doe\nTEL;TYPE=CELL:+19876543210\nEND:VCARD"
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /messages/location:
    post:
      tags: [Messages]
      summary: Send a location message
      description: Send a location with coordinates and optional name, address, and URL details. The response contains the message ID.
      operationId: sendLocation
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendLocationRequest"
            examples:
              basic:
                summary: Basic location
                value:
                  to: "1234567890@s.whatsapp.net"
                  latitude: 37.7749
                  longitude: -122.4194
              withDetails:
                summary: Location with name and address
                value:
                  to: "1234567890@g.us"
                  latitude: 51.5034
                  longitude: -0.1276
                  name: "Big Ben"
                  address: "Westminster, London SW1A 0AA, UK"
                  url: "https://maps.google.com/?q=Big+Ben"
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /messages/link:
    post:
      tags: [Messages]
      summary: Send a link preview message
      description: Send a link message with an optional preview (title, description, and JPEG thumbnail). The response contains the message ID.
      operationId: sendLink
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendLinkRequest"
            examples:
              minimal:
                summary: Minimal link
                value:
                  to: "1234567890@s.whatsapp.net"
                  text: "Check this link"
                  url: "https://example.com/article"
              withPreview:
                summary: Link with preview details
                value:
                  to: "1234567890@g.us"
                  text: "Interesting read"
                  url: "https://example.com/blog/post-1"
                  title: "A neat blog post"
                  description: "This post explains things."
                  jpegThumbnail: "/9j/4AAQSkZJRgABAQ..."
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /messages/{messageId}/reaction:
    post:
      tags: [Messages]
      summary: React to a message
      description: Send an emoji reaction to a message. Send an empty string as the reaction to remove a previous reaction.
      operationId: sendReaction
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
        - $ref: "#/components/parameters/MessageIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendReactionRequest"
            examples:
              thumbsUp:
                summary: React with thumbs up
                value:
                  to: "1234567890@s.whatsapp.net"
                  senderId: "1234567890@s.whatsapp.net"
                  reaction: "\U0001F44D"
              heart:
                summary: React with heart in a group chat
                value:
                  to: "1234567890@g.us"
                  senderId: "9876543210@s.whatsapp.net"
                  reaction: "\u2764\uFE0F"
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /messages/{messageId}/edit:
    post:
      tags: [Messages]
      summary: Edit a sent message
      description: Edit the text of a previously sent message. Only your own messages can be edited.
      operationId: editMessage
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
        - $ref: "#/components/parameters/MessageIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditMessageRequest"
            example:
              to: "1234567890@s.whatsapp.net"
              text: "Updated message"
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /messages/{messageId}/read:
    post:
      tags: [Messages]
      summary: Mark a message as read
      description: Mark a message as read with the specified receipt type.
      operationId: markAsRead
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
        - $ref: "#/components/parameters/MessageIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkAsReadRequest"
            example:
              chatId: "1234567890@s.whatsapp.net"
              senderId: "1234567890@s.whatsapp.net"
              receiptType: "read"
      responses:
        "204":
          description: Message marked as read
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /messages/{messageId}/star:
    post:
      tags: [Messages]
      summary: Star or unstar a message
      description: Star or unstar a message. Set `starred` to true to star, false to unstar.
      operationId: starMessage
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
        - $ref: "#/components/parameters/MessageIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StarMessageRequest"
            example:
              chatId: "1234567890@s.whatsapp.net"
              senderId: "1234567890@s.whatsapp.net"
              starred: true
      responses:
        "204":
          description: Message starred/unstarred
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /messages/{messageId}/pin:
    post:
      tags: [Messages]
      summary: Pin or unpin a message
      description: Pin or unpin a message in a chat. Set `pinned` to true to pin, false to unpin. Optionally specify a pin expiration duration.
      operationId: pinMessage
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
        - $ref: "#/components/parameters/MessageIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PinMessageRequest"
            examples:
              pin:
                summary: Pin a message
                value:
                  chatId: "1234567890@s.whatsapp.net"
                  senderId: "1234567890@s.whatsapp.net"
                  pinned: true
                  pinExpiration: "7d"
              unpin:
                summary: Unpin a message
                value:
                  chatId: "1234567890@s.whatsapp.net"
                  senderId: "1234567890@s.whatsapp.net"
                  pinned: false
      responses:
        "204":
          description: Message pinned/unpinned
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  securitySchemes:
    AdminAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: Admin API key for instance management endpoints (`/admin/instances/*`). Set via `auth.adminApiKey` config.

    InstanceAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: Instance API key. Resolved from the instance's own `apiKey` or the global `auth.defaultApiKey`.
  parameters:
    XInstanceId:
      name: X-Instance-Id
      in: header
      required: true
      description: The ID of the WhatsApp instance to use
      schema:
        type: string
        default: "{{instanceId}}"
...
  responses:
    Created:
      description: Resource created
      content:
        application/json:
          schema:
            type: object
            properties:
              id:
                type: string
                description: ID of the created resource
          example:
            id: "01234567890123456"
...
  schemas:
...
    SendTextRequest:
      type: object
      required: [to, text]
      properties:
        to:
          type: string
          description: Recipient JID (user or group)
        text:
          type: string
          description: Message text
        mentions:
          type: array
          items:
            type: string
          description: JIDs of mentioned users
        replyTo:
          type: string
          description: Message ID to reply to
        replyToSenderId:
          type: string
          description: Sender JID of the message being replied to
        isForwarded:
          type: boolean
          description: Mark message as forwarded
        ephemeralExpiration:
          type: string
          enum: [off, 24h, 7d, 90d]
          description: Disappearing message timer
...